pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "switchamba"
version = "0.1.0"
description = "Automatic keyboard layout switching between EN/RU/UA on Linux"
requires-python = ">=3.11"
dependencies = [
    "evdev>=1.7.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
switchamba = "switchamba.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
